pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "eventlet>=0.40.3",
    "flask>=3.1.2",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.42.0",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "python-socketio>=5.14.1",
    "werkzeug>=3.1.3",
]
